Change gallery build to use on-server images
From now on gallery images will not be committed to this git repo, so a separate copy of them must be present locally when building the site (in order to read metadata from each image).

This commit also removes a recent gallery's images, since one image had to be updated anyway for a species correction.