New in-place option only writes files if they've changed
The goal is to avoid writing the same files with every build.  To do this, it uses one build dir: "out/in-place".  Then, it uses rsync to copy, locally, only changed static files into place.  Then, all files that are built are built in bash string variables.  Then, only if the file doesn't exist or its hash is different, is it written to disk.