Lots more changes for birds pages, including overrides
Metadata overrides (currently, just bird species and sex) allow me to change the metadata used for an image if the image is wrong, or to add additional information.

The updateBirdsDbOverride.py script does interactive writing of this overrides data to the database.

The generateBirdsPage.py script does a big query to find the first, last, and best photos for each species (after applying overrides).  For each species, the subspecies and sex are also used so the first, last, and best photos for all of the following are found:
- species
- species (with subspecies)
- species (with sex)
- species (with subspecies) (and with sex)

After after doing the above big query, the generateBirdsPage.py outputs markdown content.

The generateBirdsPage.py can run a query for a specific year (or all years) and/or for a specific location (or all locations).

The generateAllBirdsPages.sh generates all the markdown pages for the queries and outputs them to the gen/static folder.  During a normal blog website build, the static markdown pages are converted to html.