Added generateBirdsPage.py
This script runs a query on the local sqlite3 bird photos database to find the first, last, and best photos for each species.  The query can be restricted to a specific year or named location (pre-computed at db insert time).

Using the query, a gallery page is constructed and output in markdown format.  The output is mostly HTML, but markdown is used for some headers and footers stuff, just like all the other static pages in the blog.