Refactor locations, add overrides for location and rating
Locations were moved to a separate table, allowing the WHERE clause that filters by location to be parameterized sanely (as opposed to modifying the SQL column name in the query depending on the location being filtered for).  A "FULL OUTER JOIN" is effectively performed in sqlite to override locations of photos and to add new locations to photos.  Ratings are overridden with a simpler LEFT OUTER JOIN.