Make gallery thumbnails much larger
This involved making a new CSS class that allows a div to be wider than its parent div.  This is the same CSS I use on the Mandelbrot gallery pages.  Since a "min-width" won't work for a wider child inside a narrower parent, we have to use multiple screen width checking @media calls to set the proper width.

To get the dates to wrap nicely across multiple lines, future galleries will have &nbsp; characters embedded directly in the formatted dates.  For all older galleries, bash regex testing is done to replace spaces with &nbsp; in date strings.

Finally, I increased the thumbnail max dimension again from 800 pixels to 900.  In the future I may want to make the thumbnails on galleries even larger, so it'll be good to have some back catalog of larger images to display.