Add "genetic algorithm" version
This uses a series of large tournaments to find better models.  The 4
best neural networks are used as the basis for half of the mutated
networks entered into the next tournament, and the other half are
randomized new neworks.

This is a smaller and simpler file than mancala-ML.html, but I suspect
it would have to run for much longer to find a strong model.