Replace mutated weights and increase overall speed
Instead of multiplying mutated weights by some factor, just replace them
entirely with new weights in what should be a better range
(Xavier/Glorot).

To increase simulation speed, greatly simplify seed-in-pit drawing and
draw things way less often.  Obviously, this should be implemented in a
multi-threaded faster language and/or done on ML hardware to get actual
decent speed.  But for this toy JavaScript program this is fine.