Fix for non-trained inputs, and experiment with slowing/skipping inputs
Board positions that result in no wins were previously not added to the
training set, but now this fix actually removes any other instances of
that board position from the previous training data.  This also does not
add, and removes and previous instances of, board positions where more
than one move share the highest winning percentange.

For slowing/skipping, the idea is to both skip learned inputs and use a
slower learning rate for nearly-learned inputs, which may:
- allow faster overall learning rate
- speed up training by skipping 'learned' inputs
- reduce overstepping by slowing 'nearly learned' inputs' learning rate
- avoid overfitting to any particular input