Port build.sh to python
It runs much faster, and should be easier to maintain and to add new
features going forward.

Changed to using MarkdownIt, which adheres to CommonMark markdown by
default, so a few small edge cases had to be tweaked here and there to
get everything rendering to HTML like it used to with the original
Markdown 1.0.1 perl script.